+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/it/SentrySpringIntegrationTest.kt
@@ -31,6 +31,7 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
@@ -145,6 +146,14 @@ class SentrySpringIntegrationTest {
                 assertThat(event.contexts.trace).isNotNull()
             }, anyOrNull())
         }
+    }
+
+    @Test
+    fun `tracing filter does not overwrite resposne status code`() {
+        val restTemplate = TestRestTemplate().withBasicAuth("user", "password")
+
+        val response = restTemplate.getForEntity("http://localhost:$port/performance", String::class.java)
+        assertThat(response.statusCode).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
     }
 
     @Test