+++ sentry-spring/src/test/kotlin/io/sentry/spring/tracing/SentryTracingFilterTest.kt
@@ -22,6 +22,7 @@ import javax.servlet.http.HttpServletRequest
 import kotlin.test.Test
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
+import kotlin.test.fail
 import org.assertj.core.api.Assertions.assertThat
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.mock.web.MockHttpServletResponse
@@ -39,7 +40,7 @@ class SentryTracingFilterTest {
             whenever(hub.options).thenReturn(SentryOptions())
         }
 
-        fun getSut(isEnabled: Boolean = true, sentryTraceHeader: String? = null): SentryTracingFilter {
+        fun getSut(isEnabled: Boolean = true, status: Int = 200, sentryTraceHeader: String? = null): SentryTracingFilter {
             request.requestURI = "/product/12"
             request.method = "POST"
             request.setAttribute(HandlerMapping.BEST_MATCHING_PATTERN_ATTRIBUTE, "/product/{id}")
@@ -47,7 +48,7 @@ class SentryTracingFilterTest {
                 request.addHeader("sentry-trace", sentryTraceHeader)
                 whenever(hub.startTransaction(any(), any<CustomSamplingContext>(), eq(true))).thenAnswer { SentryTracer(it.arguments[0] as TransactionContext, hub) }
             }
-            response.status = 200
+            response.status = status
             whenever(hub.startTransaction(any(), any(), any(), eq(true))).thenAnswer { SentryTracer(TransactionContext(it.arguments[0] as String, it.arguments[1] as String), hub) }
             whenever(hub.isEnabled).thenReturn(isEnabled)
             return SentryTracingFilter(hub, transactionNameProvider)
@@ -62,7 +63,7 @@ class SentryTracingFilterTest {
 
         filter.doFilter(fixture.request, fixture.response, fixture.chain)
 
-        verify(fixture.hub).startTransaction(eq("POST /product/12"), eq("http.server"), check<CustomSamplingContext> {
+        verify(fixture.hub).startTransaction(eq("POST /product/12"), eq("http.server"), check {
             assertNotNull(it["request"])
             assertTrue(it["request"] is HttpServletRequest)
         }, eq(true))
@@ -71,6 +72,28 @@ class SentryTracingFilterTest {
             assertThat(it.transaction).isEqualTo("POST /product/{id}")
             assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.OK)
             assertThat(it.contexts.trace!!.operation).isEqualTo("http.server")
+        })
+    }
+
+    @Test
+    fun `sets correct span status based on the response status`() {
+        val filter = fixture.getSut(status = 500)
+
+        filter.doFilter(fixture.request, fixture.response, fixture.chain)
+
+        verify(fixture.hub).captureTransaction(check {
+            assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
+        })
+    }
+
+    @Test
+    fun `does not set span status for response status that dont match predefined span statuses`() {
+        val filter = fixture.getSut(status = 302)
+
+        filter.doFilter(fixture.request, fixture.response, fixture.chain)
+
+        verify(fixture.hub).captureTransaction(check {
+            assertThat(it.contexts.trace!!.status).isNull()
         })
     }
 
@@ -108,5 +131,20 @@ class SentryTracingFilterTest {
         verify(fixture.hub).isEnabled
         verifyNoMoreInteractions(fixture.hub)
         verifyZeroInteractions(fixture.transactionNameProvider)
+    }
+
+    @Test
+    fun `sets status to internal server error when chain throws exception`() {
+        val filter = fixture.getSut()
+        whenever(fixture.chain.doFilter(any(), any())).thenThrow(RuntimeException("error"))
+
+        try {
+            filter.doFilter(fixture.request, fixture.response, fixture.chain)
+            fail("filter is expected to rethrow exception")
+        } catch (_: Exception) {
+        }
+        verify(fixture.hub).captureTransaction(check {
+            assertThat(it.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
+        })
     }
 }