+++ sentry-android-ndk/src/main/jni/sentry.c
@@ -398,5 +398,5 @@ Java_io_sentry_android_ndk_NativeModuleListLoader_nativeLoadModuleList(JNIEnv *e
 
 JNIEXPORT void JNICALL
 Java_io_sentry_android_ndk_SentryNdk_shutdown(JNIEnv *env, jclass cls) {
-    sentry_shutdown();
+    sentry_close();
 }