+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 * Fix: Activity tracing auto instrumentation for Android API < 29 (#1402)
 * Fix: use connection and read timeouts in ApacheHttpClient based transport (#1397)
 * Ref: Refactor converting HttpServletRequest to Sentry Request in Spring integration (#1387)
+* Fix: set correct transaction status for unhandled exceptions in SentryTracingFilter (#1406)
 * Fix: handle network errors in SentrySpanClientHttpRequestInterceptor (#1407)
 * Fix: set scope on transaction (#1409)
-* do not set free memory and is low memory fields when it's a NDK hard crash (#1399)
+* Fix: Do not set free memory and is low memory fields when it's a NDK hard crash (#1399)
+* Fix: Apply user from the scope to transaction (#1424)
+* Fix: Pass maxBreadcrumbs config. to sentry-native (#1425)
 
 # 4.4.0-alpha.2